requirements.txt
openai
ruamel.yaml
